Refactor Rectangle specs
It’s often helpful to group your tests inside of a `describe` block for
each method the method that they’re testing.

Since most of the tests are not testing any different behavior, just
mode examples of the same behavior, I think this is one of the cases
where it’s ok to have more than one expectation in the same `it` block.
This just helps with the readability of the file